fix: moved getTimeProps and toArray to avoid circular dependency
